refactor: modularize dependency installation and constants into separate files
